+++ R/import-standalone-pkg.R
@@ -6,7 +6,7 @@
 # ---
 # repo: Yunuuuu/standalone
 # file: standalone-pkg.R
-# last-updated: 2025-03-08
+# last-updated: 2025-03-10
 # license: https://unlicense.org
 # imports: [utils]
 # ---
@@ -16,6 +16,9 @@
 # other packages that are not listed in Imports, so use them with caution.
 
 # ## Changelog
+# 2025-03-10:
+# - Add `on_exit`
+#
 # 2025-03-08:
 # - Add `pkg_extdata`
 # - Add `defer`
@@ -75,7 +78,7 @@ pkg_extdata <- function(..., mustWork = TRUE) {
 }
 
 # Need `rlang` package, can support `quosure`
-set_exit <- function(expr, envir = parent.frame(), after = TRUE, add = TRUE) {
+on_exit <- function(expr, envir = parent.frame(), after = TRUE, add = TRUE) {
     expr <- getExportedValue("rlang", "enquo")(expr)
     defer(
         getExportedValue("rlang", "eval_tidy")(expr),